backend/appointment-service/requirements.txt
Flask==2.0.1
PyJWT==2.1.0
PyMySQL==1.1.0
Flask-Cors==4.0.1
Flask-SQLAlchemy==2.5.1
Python-dotenv==0.19.0
pika==1.2.0